requirements.txt
aniso8601==8.1.0
autopep8==1.5.4
azure-core==1.9.0
azure-storage-blob==12.6.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.19
oauthlib==3.1.0
pycodestyle==2.6.0
pycparser==2.20
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
gunicorn==20.0.4
